refactoring analysis subpackage:spectral
